# backend/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AIO ITAD API
  version: "1.0.0"
  description: |
    Backend API for IsThereAnyDeal proxy endpoints with Keycloak authentication.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /healthz:
    get:
      summary: Health check
      security: []
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: ok
  /v1/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/login:
    post:
      summary: Login and get access tokens
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/refresh:
    post:
      summary: Refresh access token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/logout:
    post:
      summary: Logout and invalidate tokens
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Logged out successfully
  /v1/auth/forgot-password:
    post:
      summary: Request password reset email
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Password reset email sent (if account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /v1/auth/resend-verification:
    post:
      summary: Resend email verification link
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationRequest"
      responses:
        "200":
          description: Verification email sent (if account exists and not verified)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /v1/auth/me:
    get:
      summary: Get current authenticated user info
      tags:
        - Authentication
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/password:
    put:
      summary: Change password
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed successfully
        "400":
          description: Validation error or incorrect current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/profile:
    put:
      summary: Update user profile
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/itad/search:
    get:
      summary: Search games
      tags:
        - Games
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyResponse"
        "400":
          description: Missing or invalid query
  /v1/itad/stores:
    get:
      summary: Get stores
      parameters:
        - name: country
          in: query
          required: false
          schema:
            type: string
            minLength: 2
            maxLength: 2
            default: DE
      responses:
        "200":
          description: Stores list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyResponse"
  /v1/itad/overview:
    get:
      summary: Price overview for multiple games
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              format: uuid
          style: form
          explode: true
          description: Repeat the id query param for multiple games
        - name: country
          in: query
          required: false
          schema:
            type: string
            minLength: 2
            maxLength: 2
            default: DE
      responses:
        "200":
          description: Overview response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyResponse"
        "400":
          description: Missing id parameter(s)
  /v1/itad/games/{gameId}:
    get:
      summary: Get combined game details (info + prices + history low)
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/Country"
      responses:
        "200":
          description: Combined response
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    $ref: "#/components/schemas/ProxyResponse"
                  prices:
                    $ref: "#/components/schemas/ProxyResponse"
                  historyLow:
                    $ref: "#/components/schemas/ProxyResponse"
  /v1/itad/games/{gameId}/info:
    get:
      summary: Get game info
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Game info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyResponse"
  /v1/itad/games/{gameId}/prices:
    get:
      summary: Get game prices
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/Country"
      responses:
        "200":
          description: Prices response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyResponse"
  /v1/itad/games/{gameId}/historylow:
    get:
      summary: Get historical low price
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/Country"
      responses:
        "200":
          description: History low response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from Keycloak
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Country:
      name: country
      in: query
      required: false
      schema:
        type: string
        minLength: 2
        maxLength: 2
        default: DE
  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          minLength: 8
          example: securePassword123
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: johndoe
        password:
          type: string
          example: securePassword123
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address to send password reset link
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address to resend verification link
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: Current password for verification
        newPassword:
          type: string
          minLength: 8
          description: New password (minimum 8 characters)
    UpdateProfileRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: New email address
        firstName:
          type: string
          description: New first name
        lastName:
          type: string
          description: New last name
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Token expiration time in seconds
        tokenType:
          type: string
          example: Bearer
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
    ProxyResponse:
      description: Raw JSON response from IsThereAnyDeal API
      type: object
      additionalProperties: true