# backend/api/internal/adapters/http/router.go
package httpapi

import (
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"

	"gamedivers.de/api/internal/adapters/http/handlers"
	authmw "gamedivers.de/api/internal/adapters/http/middleware"
)

func Router(itadh *handlers.ITADHandler, gameHandler *handlers.GameHandler, steamHandler *handlers.SteamHandler, epicHandler *handlers.EpicHandler) *chi.Mux {
func Router(itadh *handlers.ITADHandler, authh *handlers.AuthHandler, jwtMw *authmw.JWTMiddleware) *chi.Mux {
	r := chi.NewRouter()
	r.Use(middleware.RealIP)
	r.Use(middleware.RequestID)
	r.Use(middleware.Logger)
	r.Use(middleware.Recoverer)

	// CORS middleware for frontend
	r.Use(func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Access-Control-Allow-Origin", "*")
			w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
			w.Header().Set("Access-Control-Allow-Headers", "Accept, Authorization, Content-Type")

			if r.Method == "OPTIONS" {
				w.WriteHeader(http.StatusOK)
				return
			}

			next.ServeHTTP(w, r)
		})
	})

	r.Get("/healthz", func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(http.StatusOK)
		_, _ = w.Write([]byte("ok"))
	})

	r.Route("/v1", func(r chi.Router) {
		// Public auth endpoints (no authentication required)
		r.Route("/auth", func(r chi.Router) {
			r.Post("/register", authh.Register)
			r.Post("/login", authh.Login)
			r.Post("/refresh", authh.RefreshToken)
			r.Post("/logout", authh.Logout)
			r.Post("/forgot-password", authh.ForgotPassword)
			r.Post("/resend-verification", authh.ResendVerification)

			// Protected auth endpoint
			r.Group(func(r chi.Router) {
				r.Use(jwtMw.Authenticate)
				r.Get("/me", authh.GetMe)
				r.Put("/password", authh.ChangePassword)
				r.Put("/profile", authh.UpdateProfile)
			})
		})

		// Protected API endpoints (authentication required)
		r.Group(func(r chi.Router) {
			r.Use(jwtMw.Authenticate)

			// IsThereAnyDeal endpoints - provides prices from all stores including Steam
			r.Route("/itad", func(r chi.Router) {
				// Search for games
				r.Get("/search", itadh.Search)

				// Get all available stores
				r.Get("/stores", itadh.GetStores)

				// Get price overview for multiple games
				r.Get("/overview", itadh.GetOverview)

				// Game-specific endpoints
				r.Route("/games/{gameId}", func(r chi.Router) {
					// Get combined game details (info + prices + history)
					r.Get("/", itadh.GetGameDetails)

					// Get just game info
					r.Get("/info", itadh.GetGameInfo)

					// Get current prices across stores
					r.Get("/prices", itadh.GetGamePrices)

					// Get historical lowest price
					r.Get("/historylow", itadh.GetHistoricalLow)
				})
			})
		})

		// Game launch endpoints
		r.Route("/games", func(r chi.Router) {
			// Start a Steam game by app ID
			r.Post("/steam/{appid}/start", gameHandler.StartSteamGame)

			// Start an Epic Games game by app name
			r.Post("/epic/{appname}/start", gameHandler.StartEpicGame)

			// Start a GOG Galaxy game by game name
			r.Post("/gog/{gamename}/start", gameHandler.StartGOGGame)

			// Get installed/synced games
			r.Get("/installed", gameHandler.GetInstalledGames)

			// Get Steam library (placeholder)
			r.Get("/steam/library", gameHandler.GetSteamLibrary)

			// Get Epic Games library (placeholder)
			r.Get("/epic/library", gameHandler.GetEpicLibrary)

			// Get GOG Galaxy library (placeholder)
			r.Get("/gog/library", gameHandler.GetGOGLibrary)
		})

		// Steam authentication and library
		r.Route("/steam", func(r chi.Router) {
			// Initiate Steam login
			r.Get("/login", steamHandler.LoginRedirect)

			// Steam OpenID callback
			r.Get("/callback", steamHandler.Callback)

			// Get Steam library for authenticated user
			r.Get("/library", steamHandler.GetLibrary)

			// Sync Steam library to database
			r.Post("/sync", steamHandler.SyncLibrary)
		})

		// Epic Games authentication and library
		r.Route("/epic", func(r chi.Router) {
			// Initiate Epic Games login
			r.Get("/login", epicHandler.LoginRedirect)

			// Epic OAuth callback
			r.Get("/callback", epicHandler.Callback)

			// Get Epic Games library for authenticated user
			r.Get("/library", epicHandler.GetLibrary)

			// Sync Epic Games library to database
			r.Post("/sync", epicHandler.SyncLibrary)
		})
	})

	return r
}